Deprecate old trait names
