+++ Traits/Gedmo/PersonalTranslatable.php
@@ -11,55 +11,20 @@
 
 namespace Sonata\TranslationBundle\Traits\Gedmo;
 
-use Sonata\TranslationBundle\Model\Gedmo\AbstractPersonalTranslation;
-use Sonata\TranslationBundle\Traits\Translatable;
+@trigger_error(
+    'The '.__NAMESPACE__.'\PersonalTranslatable class is deprecated since version 2.x and will be removed in 3.0.'.
+    'Use the '.__NAMESPACE__.'\PersonalTranslatableTrait class instead.',
+    E_USER_DEPRECATED
+);
 
 /**
  * If you don't want to use trait, you can extend AbstractPersonalTranslatable instead.
  *
  * @author Nicolas Bastien <nbastien.pro@gmail.com>
+ *
+ * @deprecated since version 2.x and will be removed in 3.0
  */
 trait PersonalTranslatable
 {
-    use Translatable;
-
-    /**
-     * @return ArrayCollection|AbstractPersonalTranslation[]
-     */
-    public function getTranslations()
-    {
-        return $this->translations;
-    }
-
-    /**
-     * @param $field
-     * @param $locale
-     *
-     * @return null|string
-     */
-    public function getTranslation($field, $locale)
-    {
-        foreach ($this->getTranslations() as $translation) {
-            if (strcmp($translation->getField(), $field) === 0 && strcmp($translation->getLocale(), $locale) === 0) {
-                return $translation->getContent();
-            }
-        }
-
-        return;
-    }
-
-    /**
-     * @param AbstractPersonalTranslation $translation
-     *
-     * @return $this
-     */
-    public function addTranslation(AbstractPersonalTranslation $translation)
-    {
-        if (!$this->translations->contains($translation)) {
-            $translation->setObject($this);
-            $this->translations->add($translation);
-        }
-
-        return $this;
-    }
+    use PersonalTranslatableTrait;
 }
